random walk for cards and suits
This code answers the following questions:
You have a deck with N different playing cards, equally distributed amongst M suits. You draw all cards without putting any back in the deck. After drawing the first card, you compare the suit of each subsequent card drawn with the suit of the card drawn immediately before. If the suits match, you get a point. Otherwise, you get no points. The total number of points is P at the end of the process.
What is the mean of P for different combinations of N and M?
What is the standard deviation of P for different combinations of N and M?
What is the conditional probability that P is greater than 12 given that it is greater than 6 for different combinations of N and M?